feature/EPMLCID-15088 covering missing queue items and stall jobs

+++ PortfolioEngineCore/PortfolioEngineCore/Services/QueueManager.cs
@@ -175,7 +175,7 @@ namespace PortfolioEngineCore
             return bHandled;
         }
 
-        public long ManageTimedJobs()
+        public int ManageTimedJobs()
         {
          
             try
